+++ XYUI/src/XYUI.Avalonia/Controls/XYToggleSwitch.cs
@@ -1,8 +1,8 @@
-using Avalonia.Controls;
+using Avalonia.Controls.Primitives;
 
 namespace XYUI.Avalonia.Controls;
 
-public sealed class XYToggleSwitch : ToggleSwitch
+public sealed class XYToggleSwitch : ToggleButton
 {
     public XYToggleSwitch() => Classes.Add("xy-toggle-switch");
 }